requirements.txt
asgiref==3.2.3
astroid==2.3.3
bandit==1.6.2
cffi==1.14.0
colorama==0.4.3
cryptography==2.9.2
Django==3.0.4
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
gitdb==4.0.4
GitPython==3.1.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy==0.720
mypy-extensions==0.4.3
pbr==5.4.5
psycopg2==2.8.5
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.0.15
pylint-plugin-utils==0.6
pytz==2019.3
PyYAML==5.3.1
six==1.14.0
smmap==3.0.2
sqlparse==0.3.1
stevedore==1.32.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
wrapt==1.11.2
